some careful type casts to persuade the compiler to be silent
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@1211 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ControlPanel.cxx
@@ -297,9 +297,9 @@ void			ControlPanel::resize()
     messageFont = TextureFont::getTextureFont(TextureFont::Fixed, true);
   messageFont.setSize(fontSize, fontSize);
 
-  maxLines = messageAreaPixels[3] / fontSize;
+  maxLines = int(messageAreaPixels[3] / fontSize);
 
-	   // note that we've been resized at least once
+  // note that we've been resized at least once
   resized = true;
   expose();
 }
@@ -368,7 +368,7 @@ void			ControlPanel::addMessage(const std::string& line,
 {
   ControlPanelMessage item(line, color);
 
-  if (messages.size() < maxLines * maxScrollPages) {
+  if ((int)messages.size() < maxLines * maxScrollPages) {
     // not full yet so just append it
     messages.push_back(item);
   }
